backend/requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==3.2.20
django-cors-headers==4.2.0
djangorestframework==3.14.0
filelock==3.12.4
platformdirs==3.11.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
virtualenv==20.24.5
